+++ addon/components/flexberry-tab-bar.js
@@ -1,6 +1,7 @@
 import Component from '@ember/component';
-import { get, set } from '@ember/object';
-import { A, isArray } from '@ember/array';
+import { get, set, computed } from '@ember/object';
+import { A } from '@ember/array';
+import $ from "jquery"
 
 /**
   Component's CSS-classes names.
@@ -32,6 +33,8 @@ const flexberryClassNames = {
 export default Component.extend({
   classNames: ['ui', 'tabular', 'menu', flexberryClassNamesPrefix],
 
+  classNameBindings: ['isOverflowedTabs:overflowed'],
+
   /**
     Reference to component's CSS-classes names.
     Must be also a component's instance property to be available from component's .hbs template.
@@ -49,10 +52,153 @@ export default Component.extend({
   items: null,
 
   /**
+   * Contains items that will be displayed in tab bar
+   * @property tabs
+   * @type {Array}
+   */
+  tabs: computed('items.{[],@each.active}', function () {
+    let active = false;
+    let items = this.get('items') || A();
+    let result = A();
+
+    items.forEach((item) => {
+      let itemIsActive = get(item, 'active');
+      if (itemIsActive) {
+        if (!active) {
+          active = true;
+
+          let itemClass = get(item, 'class') || '';
+          let regEx = /\sactive(\s|$)/;
+          if (!regEx.test(itemClass)) {
+            itemClass += ' active';
+          }
+
+          set(item, 'class', itemClass);
+          this.set('prevTab', item.selector);
+          $.tab('change tab', item.selector);
+        }
+      }
+
+      set(item, 'active', false);
+
+      if (get(item, 'iconClass')) {
+        set(item, '_hasIcon', true);
+      } else {
+        set(item, '_hasIcon', false);
+      }
+
+      result.pushObject(item);
+    });
+
+    return result;
+  }),
+
+  /**
    * Contains name of previous data-tab
    * @property prevTab
    */
   prevTab: undefined,
+
+  /**
+   * String with dropdown selector for working with jQuery
+   * @property dropdownDomString
+   */
+  navDropdownDomString: '.ui.compact.pointing.top.right.dropdown.link.item',
+
+  isOverflowedTabs: false,
+
+  /**
+   * Checks if sum of tabs width is greater than tab container.
+   * If true - dropdown becomes visible.
+   * If false - dropdown dissapears.
+   * @method setDropdownVisibility
+   */
+  setNavDropdownVisibility: function () {
+    const tabContainer = document.querySelector(
+      '.ui.tabular.menu.flexberry-tab-bar'
+    );
+    const tab = document.querySelector('.flexberry-tab-bar-tab.tab.item');
+
+    if (tab != null && tab.clientWidth * this.items.length > tabContainer.clientWidth) {
+      this.$(this.navDropdownDomString).show();
+    } else {
+      this.$(this.navDropdownDomString).hide();
+    }
+  },
+
+  /**
+   * Enables scroll wheel for dragscroll.
+   * Based on code by @miorel + @pieterv of Facebook.
+	 * github.com/facebook/fixed-data-table/blob/master/src/vendor_upstream/dom/normalizeWheel.js
+   * @method normalizeWheel
+   */
+  normalizeWheel: function (event) {
+    const pixelStep = 10;
+    const lineHeight = 40;
+    const pageHeight = 800;
+    let sX = 0;
+    let sY = 0;
+    let pX = 0;
+    let pY = 0;
+
+    // Legacy.
+    if ('detail' in event) {
+      sY = event.detail;
+    } else if ('wheelDelta' in event) {
+      sY = event.wheelDelta / -120;
+    } else if ('wheelDeltaY' in event) {
+      sY = event.wheelDeltaY / -120;
+    }
+
+    if ('wheelDeltaX' in event) {
+      sX = event.wheelDeltaX / -120;
+    }
+
+    // Side scrolling on FF with DOMMouseScroll.
+    if ('axis' in event && event.axis === event.HORIZONTAL_AXIS) {
+      sX = sY;
+      sY = 0;
+    }
+
+    // Calculate.
+    pX = sX * pixelStep;
+    pY = sY * pixelStep;
+
+    if ('deltaY' in event) {
+      pY = event.deltaY;
+    }
+
+    if ('deltaX' in event) {
+      pX = event.deltaX;
+    }
+
+    if ((pX || pY) && event.deltaMode) {
+      if (event.deltaMode === 1) {
+        pX *= lineHeight;
+        pY *= lineHeight;
+      } else {
+        pX *= pageHeight;
+        pY *= pageHeight;
+      }
+    }
+
+    // Fallback if spin cannot be determined.
+    if (pX && !sX) {
+      sX = pX < 1 ? -1 : 1;
+    }
+
+    if (pY && !sY) {
+      sY = pY < 1 ? -1 : 1;
+    }
+
+    // Return.
+    return {
+      spinX: sX,
+      spinY: sY,
+      pixelX: pX,
+      pixelY: pY,
+    };
+  },
 
   actions: {
     /**
@@ -63,7 +209,7 @@ export default Component.extend({
     */
     change(currentTab, event) {
       let prevTab = this.get('prevTab');
-      let tabName = currentTab.selector;
+      let tabName = currentTab;
       let changed = false;
 
       if (prevTab !== tabName) {
@@ -86,34 +232,11 @@ export default Component.extend({
         tabName: tabName,
         prevTab: prevTab,
         changed: changed,
-        originalEvent: event
+        originalEvent: event,
       };
 
-      this.get('change')(e);
-    }
-  },
-
-  init() {
-    this._super(...arguments);
-
-    let items = this.get('items');
-    if (!isArray(items)) {
-      items = A(items)
-      this.set('items', items);
-    }
-
-    let active = false;
-    items.forEach((item) => {
-      let itemIsActive = get(item, 'active');
-      if (itemIsActive && !active) {
-        active = true;
-
-        this.set('prevTab', item.selector);
-        this.$().tab('change tab', item.selector);
-      } else {
-        set(item, 'active', false);
-      }
-    }, this);
+      this.get('change', e);
+    },
   },
 
   /**
@@ -124,6 +247,68 @@ export default Component.extend({
 
     // initialize semantic ui tabs
     this.$('.item').tab();
+
+    if (this.get('isOverflowedTabs')) {
+      // Dragscroll inplementation for tabs
+      const slider = document.querySelector('.dragscroll');
+      let isDown = false;
+      let startX;
+      let scrollLeft;
+
+      slider.addEventListener('mousedown', (e) => {
+        isDown = true;
+        slider.classList.add('active');
+        startX = e.pageX - slider.offsetLeft;
+        scrollLeft = slider.scrollLeft;
+      });
+
+      slider.addEventListener('mouseleave', () => {
+        isDown = false;
+        slider.classList.remove('active');
+      });
+
+      slider.addEventListener('mouseup', () => {
+        isDown = false;
+        slider.classList.remove('active');
+      });
+
+      slider.addEventListener('mousemove', (e) => {
+        if (!isDown) {
+          return false;
+        } else {
+          e.preventDefault();
+          const x = e.pageX - slider.offsetLeft;
+          const walk = (x - startX) * 1.5;
+          slider.scrollLeft = scrollLeft - walk;
+        }
+      });
+
+      slider.addEventListener('wheel', (e) => {
+
+        // Prevent default.
+        e.preventDefault();
+        e.stopPropagation();
+
+        // Stop link scroll.
+        this.$('body').stop();
+
+        // Calculate delta, direction.
+        const n = this.normalizeWheel(e);
+        const x = n.pixelX !== 0 ? n.pixelX : n.pixelY;
+        const delta = Math.min(Math.abs(x), 150);
+        const direction = x > 0 ? 1 : -1;
+
+        // Scroll page.
+        this.$('.dragscroll').scrollLeft(this.$('.dragscroll').scrollLeft() + delta * direction);
+      });
+
+      // Dropdown visibility implementation
+      window.addEventListener('resize', () => {
+        this.setNavDropdownVisibility();
+      });
+
+      this.setNavDropdownVisibility();
+    }
   },
 
   /**
@@ -134,6 +319,17 @@ export default Component.extend({
 
     // Initialize possibly added new tabs.
     this.$('.item').tab();
+
+    if (this.get('isOverflowedTabs')) {
+      // Inititalize semantic ui dropdown (hidden by default)
+      this.$(this.navDropdownDomString).dropdown({
+        transition: 'drop',
+        action: 'activate',
+        onChange(newTab) {
+          $.tab('change tab', newTab);
+        },
+      });
+    }
   },
 
   /**
@@ -144,7 +340,7 @@ export default Component.extend({
 
     // destroy semantic ui tabs
     this.$('.item').tab('destroy');
-  }
+  },
 
   /**
     Component's action invoking when tab was clicked and it's 'active' state changed.