+++ app/controllers/editrecord-dialog.js
@@ -1,0 +1,2 @@
+import EditrecordDialogController from 'ember-flexberry/controllers/editrecord-dialog';
+export default EditrecordDialogController;